+++ tests/testthat/test_measures_manipulation.R
@@ -4,13 +4,15 @@ context("Measures manipulation")
 library("sentometrics")
 library("quanteda")
 
+set.seed(123)
+
 # corpus, lexicon and aggregation control creation
 data("usnews")
 corpus <- quanteda::corpus_sample(sento_corpus(corpusdf = usnews), size = 1000)
 
 data("list_lexicons")
 lex <- list_lexicons[c("GI_en", "LM_en")]
-ctr <- ctr_agg(howWithin = "tf-idf", howDocs = "proportional", by = "week", lag = 21, do.ignoreZeros = TRUE,
+ctr <- ctr_agg(howWithin = "proportionalPol", howDocs = "proportional", by = "week", lag = 21, do.ignoreZeros = TRUE,
                howTime = c("almon", "beta", "linear"), ordersAlm = 1:3, do.inverseAlm = TRUE, aBeta = 1:3, bBeta = 1:2)
 
 sentMeas <- sento_measures(corpus, lex, ctr)