+++ tests/testthat/test_sentiment_computation.R
@@ -18,15 +18,15 @@ lexSplit <- setup_lexicons(list_lexicons[c("GI_en", "LM_en", "HENRY_en")], do.sp
 
 sentimentList <- list(
   s1 = compute_sentiment(quanteda::texts(corpus), lex, how = "counts"),
-  s2 = compute_sentiment(quanteda::texts(corpus), lex[names(lex) != "valence"], how = "tf-idf"),
+  s2 = compute_sentiment(quanteda::texts(corpus), lex[names(lex) != "valence"], how = "counts"),
   s3 = compute_sentiment(quanteda::texts(corpus), lex, how = "proportional"),
-  s4 = compute_sentiment(quanteda::texts(corpus), lex, how = "proportionalPol"),
+  s4 = compute_sentiment(quanteda::texts(corpus), lex, how = "proportionalPol", nCore = 2),
   s5 = compute_sentiment(quanteda::corpus(usnews[1:250, "texts"]), lex, how = "counts"),
   s6 = compute_sentiment(quanteda::corpus(usnews[1:250, c("texts", "wsj", "economy")], text_field = "texts"),
                          lex, how = "counts"),
   s7 = compute_sentiment(corpus, lex, how = "counts"),
   s8 = compute_sentiment(quanteda::texts(corpus), lexSplit, how = "counts"),
-  s9 = compute_sentiment(quanteda::texts(corpus), lex[names(lex) != "valence"], how = "tf-idf", nCore = c(2, 2))
+  s9 = compute_sentiment(quanteda::texts(corpus), lex, how = "proportionalPol", nCore = 1)
 )
 
 # compute_sentiment
@@ -34,16 +34,15 @@ test_that("Agreement between sentiment scores across input objects", {
   expect_true(all(unlist(lapply(sentimentList, function(s) nrow(s$sentiment) == 250))))
   expect_true(all(unlist(lapply(sentimentList, function(s) all(s$sentiment$word_count
                                                                == sentimentList$s1$sentiment$word_count)))))
-  expect_true(all(sentimentList$s8$sentiment[, c("GI_en_POS", "LM_en_POS", "HENRY_en_POS")] >= 0))
-  expect_true(all(sentimentList$s8$sentiment[, c("GI_en_NEG", "LM_en_NEG", "HENRY_en_NEG")] <= 0))
-  expect_true(all(sentimentList$s2$sentiment == sentimentList$s9$sentiment))
-  expect_equivalent(sentimentList$s1$sentiment[, c("GI_en", "LM_en", "HENRY_en")],
+  expect_true(all(sentimentList$s8[, c("GI_en_POS", "LM_en_POS", "HENRY_en_POS")] >= 0))
+  expect_true(all(sentimentList$s8[, c("GI_en_NEG", "LM_en_NEG", "HENRY_en_NEG")] <= 0))
+  expect_true(all(sentimentList$s4 == sentimentList$s9))
+  expect_equivalent(sentimentList$s1[, c("GI_en", "LM_en", "HENRY_en")],
                     sentimentList$s5$sentiment[, c("GI_en", "LM_en", "HENRY_en")])
   expect_equivalent(sentimentList$s6$sentiment[, -c(1:2)],
                     sentimentList$s7$sentiment[, colnames(sentimentList$s6$sentiment)[-c(1:2)], with = FALSE])
-  expect_error(compute_sentiment(quanteda::texts(corpus), lex, how = "tf-idf"))
   expect_error(compute_sentiment(quanteda::texts(corpus), lex, how = "notAnOption"))
-  expect_warning(compute_sentiment(quanteda::texts(corpus), lex[names(lex) != "valence"], how = "tf-idf", nCore = c(-3, 0)))
+  expect_warning(compute_sentiment(quanteda::texts(corpus), lex[names(lex) != "valence"], how = "counts", nCore = -1))
 })
 
 # setup_lexicons